Cast SD3 sample draws to fp32 before metrics -- cdist/torchmetrics have no bf16 CUDA kernel
flow_sd3.py's smoke-test run got through the full training loop for the
first time this session (50 steps, ~0.95 it/s) and then hit the first
eval-time bug: evaluate_sd3's condMMD calc calls g1.mmd_rff -> torch.cdist,
which has no CUDA kernel for bf16 ("cdist_cuda" not implemented for
'BFloat16'). sample_flow_sd3's draws come back in net.dtype (bf16 in this
run) since decode_latent casts to self.dtype. Fixed at the source in both
evaluate_sd3 (real/flow_sd3.py) and _draw_samples (real/eval.py, used by
the full Kodak eval protocol's MS-SSIM/LPIPS/FID -- same bf16-unsupported
shape of bug, not yet hit but certain to be next).

+++ real/eval.py
@@ -59,7 +59,12 @@ def _draw_samples(net, codec, image, lam, n_draws, n_steps, backbone):
     from real.flow_sd3 import sample_flow_sd3
 
     cond_latent = net.encode_pixels(codec.decode(y_hat))
-    return torch.stack([sample_flow_sd3(net, cond_latent, lam, n_steps) for _ in range(n_draws)])
+    # .float(): sd3's draws come back in net.dtype (e.g. bf16); torchmetrics'
+    # MS-SSIM/LPIPS/FID and torch.cdist-based metrics below have no bf16 CUDA
+    # kernel support, so this needs fp32 regardless of what dtype training ran in.
+    return torch.stack(
+        [sample_flow_sd3(net, cond_latent, lam, n_steps) for _ in range(n_draws)]
+    ).float()
 
 
 @torch.no_grad()
